+++ src/lib/PnP.Framework/Provisioning/Providers/Xml/Resolvers/V201807/SiteCollectionsAndSitesFromSchemaToModelTypeResolver.cs
@@ -34,6 +34,8 @@ namespace PnP.Framework.Provisioning.Providers.Xml.Resolvers.V201807
             var teamSiteType = Type.GetType(teamSiteTypeName, true);
             var teamSiteNoGroupTypeName = $"{PnPSerializationScope.Current?.BaseSchemaNamespace}.TeamSiteNoGroup, {PnPSerializationScope.Current?.BaseSchemaAssemblyName}";
             var teamSiteNoGroupType = Type.GetType(teamSiteNoGroupTypeName, true);
+            var classicSiteTypeName = $"{PnPSerializationScope.Current?.BaseSchemaNamespace}.ClassicSite, {PnPSerializationScope.Current?.BaseSchemaAssemblyName}";
+            var classicSiteType = Type.GetType(classicSiteTypeName, true);
             var teamSubSiteNoGroupTypeName = $"{PnPSerializationScope.Current?.BaseSchemaNamespace}.TeamSubSiteNoGroup, {PnPSerializationScope.Current?.BaseSchemaAssemblyName}";
             var teamSubSiteNoGroupType = Type.GetType(teamSubSiteNoGroupTypeName, true);
 
@@ -61,6 +63,10 @@ namespace PnP.Framework.Provisioning.Providers.Xml.Resolvers.V201807
                     else if (i.GetType().Name == teamSiteNoGroupType.Name)
                     {
                         targetItem = new Model.TeamNoGroupSiteCollection();
+                    }
+                    else if (i.GetType().Name == classicSiteType.Name)
+                    {
+                        targetItem = new Model.ClassicSiteCollection();
                     }
                     else if (i.GetType().Name == teamSubSiteNoGroupType.Name)
                     {